+++ NEWS.md
@@ -1,3 +1,5 @@
+# erahumed (development version)
+
 # erahumed 0.11.0
 
 ### API changes