+++ NEWS.md
@@ -5,6 +5,9 @@
 * Fixed functionality of `assert_data.frame(extends = FALSE)` (the
 `extends` argument was being bypassed).
 
+* Removed low-level creator for `erahumed_simulation` class objects (*i.e.* 
+`new_erahumed_simulation()`).
+
 ### Testing
 
 * Added snapshot tests for plot methods of ERAHUMED layers (#149).