+++ NEWS.md
@@ -18,6 +18,8 @@
 * Added basic server tests for all shiny modules (#150).
 * Added basic UI no-error tests for all shiny modules.
 * Testing of `run_simulation()`.
+* Test that setting seed in `setup_hbp()` ensures reproducibility, and that 
+different seeds give rise to different outputs.
 * Testing corner cases of assertions.
 * Testing of `print()` and `summary()` S3 methods for `erahumed_simulation` and 
 `erahumed_simulation_layer` S3 classes.